% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell-statistics.R
\name{cell_intensity.multiplot}
\alias{cell_intensity.multiplot}
\title{Cell multiplot}
\usage{
cell_intensity.multiplot(samples = batch.samples)
}
\arguments{
\item{vector}{with directories}
}
\description{
Basic plot of cell intensity distribution
}
\examples{
cell_intensity.multiplot()
}
\keyword{cell}

