% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell-statistics.R
\name{cell_intensity.plot}
\alias{cell_intensity.plot}
\title{cell statistics}
\usage{
cell_intensity.plot(data, kernel = TRUE, sample.name = NULL, hide = FALSE,
  graphTitle = "Cell intensity distribution")
}
\arguments{
\item{data}{frame}
}
\description{
Basic plot of cell intensity distribution
}
\examples{
cell_intensity.plot()
}
\keyword{cell}

