% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-Environment.R
\name{isbatch_setEnvironment}
\alias{isbatch_setEnvironment}
\title{Set variables to start an analysis}
\usage{
isbatch_setEnvironment(experiment = "./")
}
\arguments{
\item{Experiment}{path}
}
\description{
This function set the basic variables for convenience
}
\examples{
isbatch_setEnvironment()
}
\keyword{experiment}

