% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeSample}
\alias{mergeSample}
\title{Merge tables in Samples}
\usage{
mergeSample(sample = ".", tableName = NULL)
}
\arguments{
\item{Sample}{folder path}
}
\description{
This function returns a single data frame containing data from all indicated tables.
}
\examples{
mergeSample()
}
\keyword{sample}

